requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
